+++ CHANGELOG.md
@@ -12,8 +12,8 @@ AssemblyAdapter:
 * bug: Fixed an issue where IndexOutOfBoundsException might be thrown in the getSpanSize() method of ItemActor when using AssemblyGridLayoutManager
 * remove: Remove AssemblyAdapter's getItemFactoryList(), getItemFactoryCount(), getData(int) methods
 * remove: Remove AssemblyAdapter's getHeaderItemList(), getHeaderData(), getFooterItemList(), getFooterData() methods
-* change: Rename AssemblyAdapter's getHeaderItemCount() to getHeaderEnabledItemCount()
-* change: Rename AssemblyAdapter's getFooterItemCount() to getFooterEnabledItemCount()
+* change: Rename AssemblyAdapter's getHeaderItemCount() to getHeaderCount()
+* change: Rename AssemblyAdapter's getFooterItemCount() to getFooterCount()
 * change: Rename AssemblyAdapter's getMoreItemHolder() to getMoreItem()
 * new: AssemblyAdapter's adds getHeaderItemManager() and getFooterItemManager() methods
 * remove: Remove ItemHolder's getItemStorage(), getPosition(), setPosition(int) methods